Changing place of roorm code and adding fonctionnalities for rooprint
    o rooprint : add an option 'extension' to choose the output extension
    o rooprint : add an option 'directory' to place outputs on a subdirectory
    o rooprint : modify option 'output' to work only on ps or pdf
    o like for roomkdir_mkdir or roocp_copy, place the entire code
      for the removing in roorm_delete
    o some modifications on utils

+++ python/roorm_delete.py
@@ -6,11 +6,22 @@ from utils import *
 import os
 import ROOT
 
-def roorm_delete(root_file,path):
-    """Remove the object corresponding to the path from a root file"""
-    if path != []:
-        chg_dir(root_file,path[:-1])
-        ROOT.gDirectory.Delete(path[-1]+";*")
-        # It doesn't work without the +";*" for cycles
-    else:
-        os.system("rm {}".format(root_file.GetName()))
+def roorm_delete(root_file,path,opt_dict):
+    """Remove the object corresponding to the path from a root file
+    depending on options"""
+    do_remove = True
+    if not opt_dict['force']:
+        if path != []:
+            answer = raw_input("Are you sure to remove '{0}' from '{1}' ? (y/n) : " \
+                                   .format("/".join(path),file_name))
+        else:
+            answer = raw_input("Are you sure to remove '{0}' ? (y/n) : " \
+                                   .format(file_name))
+        if answer.lower() != 'y':
+            do_remove = False
+    if do_remove:
+        if path != []:
+            chg_dir(root_file,path[:-1])
+            ROOT.gDirectory.Delete(path[-1]+";*")
+        else:
+            os.system("rm {}".format(root_file.GetName()))
